make sure we use the parsed key fields during unmarshling
